Discrete zodiac colors

Discrete zodiac colors

scale_fill_zodiac(..., use_factor_order = FALSE)

Arguments

...
Extra parameters to pass to ggplot2::discrete_scale() when use_factor_order is FALSE (default).

use_factor_order
A logical. If TRUE, this will assign colors based on how they match with the zodiac sign. If FALSE (default) then the colors will be used without looking for zodiac sign values.

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

library(ggplot2)

## Use colors that match with each zodiac sign
## sample_dates comes with the package to demonstrate adding zodiac symbols to dates in 2020, all dates from 2000-01-01 to 2000-12-31 are present
p <-
  sample_dates %>%
  ## Sample 100 dates from 2020 to count
  sample_n(100) %>%
  mutate(zodiac_sign = zodiac_sign(date)) %>%
  ggplot() +
  aes(
    x = zodiac_sign,
    fill = zodiac_sign
  ) +
  geom_bar() +
  scale_fill_zodiac(use_factor_order = TRUE)

#> Error: Problem with `mutate()` column `zodiac_sign`.
#> ℹ `zodiac_sign = zodiac_sign(date)`.
#> ✖ there is no package called ‘lubridate’

## Use the colors in the palette without zodiac signs
p2 <-
  iris %>%
  ggplot() +
  aes(
    x = Species,
    fill = Species
  ) +
  geom_bar() +
  scale_fill_zodiac()

#> Error in zodiac_color_palette(): object 'zodiac_signs_colors' not found